requirements.txt
ez-yaml==1.1.0
PyYAML==6.0
typing==3.7.4.3
